+++ tasks/config.js
@@ -31,7 +31,7 @@ module.exports = {
     ` * ${name} v${release}`,
     ' *',
     ' * A Javascript View library for building web and hybrid mobile apps.',
-    ' * Copyright (c) 2017 Jclo <jclo@mobilabs.fr> (http://www.mobilabs.fr).',
+    ' * Copyright (c) 2018 Jclo <jclo@mobilabs.fr> (http://www.mobilabs.fr).',
     ' * Released under the MIT license. You may obtain a copy of the License',
     ' * at: http://www.opensource.org/licenses/mit-license.php).',
     ' */',