+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ### HEAD
 
+### 0.0.8 (January 18, 2018)
+
+  * Added the methods $().select, $().selectChild, $().parent, $().firstParent,
+  * Added the methods $().clone, $().insertChildBefore(), $().removeChild, $().replaceChild,
+  * Added the methods $().children, $().childrenIndex, $().getRect(),
+  * Added the methods $().addClasses, $().removeClasses, $().hasClass,
+  * ...,
+
+
 ### 0.0.7 (August 4, 2017)
 
   * Fixed a few typos in the README,